feat: change data type

+++ contracts/protocol/payload-delivery/app-gateway/DeliveryHelper.sol
@@ -16,7 +16,7 @@ contract DeliveryHelper is BatchAsync {
     function initialize(
         address addressResolver_,
         address owner_,
-        uint256 bidTimeout_
+        uint128 bidTimeout_
     ) public reinitializer(1) {
         _setAddressResolver(addressResolver_);
         bidTimeout = bidTimeout_;
